# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "easyagent控制管理平台",
        "title": "easyagent",
        "termsOfService": "https://github.com/mufanh/easyagent",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/api/sessions": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "获取agent连接列表",
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/model.SessionInfo"
                        }
                    },
                    "400": {
                        "description": "请求错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.Error"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.Error"
                        }
                    }
                }
            }
        },
        "/api/sessions/{id}": {
            "delete": {
                "produces": [
                    "application/json"
                ],
                "summary": "关闭agent连接",
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "请求错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.Error"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.Error"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "errcode.Error": {
            "type": "object"
        },
        "model.SessionInfo": {
            "type": "object"
        }
    }
}